requirements.txt
pandas==1.4.3
PyYAML==6.0
requests==2.28.1
setuptools==65.6.3
